Add migration for export fix
* Review fixes
